requirements.txt
Django
python-dotenv
nanoid
pytelegrambotapi